Bug 1419440 - Update webrender to commit b7b07562fda338fcb2faff66ce01aafb6235fbcf. r=mstange
MozReview-Commit-ID: F1WeQbL6Bhg

UltraBlame original commit: 4bd54453fd2e845387d025804b12342614f28a4e

+++ gfx/webrender_api/src/display_item.rs
@@ -60,7 +60,9 @@ bitflags! {
 
 
 
-pub type ItemTag = (u64, u8);
+
+
+pub type ItemTag = (u64, u16);
 
 #[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
 pub struct DisplayItem {
